master/requirements.txt
pathway>=0.7.0
PyMuPDF==1.23.7
pathlib
pandas>=2.0.0
bottleneck>=1.3.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
tenacity>=8.0.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
openai>=1.0.0
torch>=2.0.0
colorama>=0.4.6
tabulate>=0.9.0
anthropic
arxiv>=1.4.8
requests>=2.31.0
tqdm>=4.66.1
PyPDF2>=3.0.1
termcolor>=2.3.0
